refactor: Relocate example pages to main.dart
This commit moves the following example page implementations from their individual files into `example/lib/main.dart`:
- `BlocLoadingDemoPage`
- `BlocOnboardingDemoPage`
- `BlocResponsiveDemoPage`
- `ConnectivityDemoPage`
- `HomePage`
- `SessionDemoPage`
- `WsDatabaseUserDemoPage`

The corresponding individual files have been deleted.
The `GatewayWsDatabase` abstract class has also been moved to `main.dart`.
